+++ src/Artatawe/Data/Painting.java
@@ -1,13 +1,13 @@
 package Artatawe.Data;
 
 import java.util.Date;
-import javafx.scene.image.Image;
+
 /**
  * @author Charlie Daley
  */
 public class Painting extends Artwork{
 
-    public Painting(String name, String description, Image photo, int year, int reservedPrice, int allowedBids, Date dateTime, int width, int height){
+    public Painting(String name, String description, Image photo, int year, int reservedPrice, int allowedBids, Date dateTime, double width, double height){
         super(name, description, photo, year, reservedPrice, allowedBids, dateTime, width, height);
